--- app/model/Todo.scala
@@ -1,13 +1,0 @@
-
-package model
-
-// Topページのviewvalue
-case class ViewValueTodo(
-
-  title:  String,
-  cssSrc: Seq[String],
-  jsSrc:  Seq[String],
-
-  todoList: Seq[TodoWithCategory]
-
-) extends ViewValueCommon
