+++ app/views/category/Category.scala.html
@@ -16,18 +16,18 @@
         <th>@category.v.name</th>
         <td>@category.v.slug</td>
         <td><font color=@category.v.color.name>■</font></td>
-        <td><a href="@{routes.CategoryController.editHome(category.id.toInt)}"><span>編集</span></a></td>
+        <td><a href="@{routes.CategoryController.getCategory(category.id.toInt)}"><span>編集</span></a></td>
         <td>
-          @helper.form( action = helper.CSRF( routes.CategoryController.delete(category.id.toInt ) ) ){
+          @helper.form( action = helper.CSRF( routes.CategoryController.deleteCategory(category.id.toInt ) ) ){
             <input type="submit" value="削除">
           }
         </td>
       <tr>
     }
     <tr>
-      <th><a href="@{routes.CategoryController.addHome}"><span>追加</span></a></th>
+      <th><a href="@{routes.CategoryController.addCategoryHome}"><span>追加</span></a></th>
     </tr>
   </table>
-  <a href="@{routes.HomeController.getTodoList}"><span>戻る</span></a>
+  <a href="@{routes.TodoController.getTodoList}"><span>戻る</span></a>
 
 }
