+++ app/views/category/ErrorEdit.scala.html
@@ -1,3 +1,3 @@
 
 <h2>更新に失敗しました。</h2>
-<a href="@{routes.CategoryController.getList}"><span>一覧に戻る</span></a>
+<a href="@{routes.CategoryController.getCategoryList}"><span>一覧に戻る</span></a>
