+++ app/views/todo/Edit.scala.html
@@ -5,8 +5,9 @@ categorySeq: Seq[lib.model.TodoCategory#EmbeddedId],
 filledEditTodoForm: Form[forms.EditTodoData]
 )(implicit request: RequestHeader, messagesProvider: MessagesProvider)
 
+
 @common.Default(vv){
-   @helper.form( action = helper.CSRF( routes.HomeController.editTodo ) ){
+   @helper.form( action = helper.CSRF( routes.TodoController.editTodo ) ){
       @helper.inputText(filledEditTodoForm("title"), '_label -> "タイトル")
       @helper.textarea(filledEditTodoForm("body"), '_label -> "詳細")
       @helper.select(field = filledEditTodoForm("status"), options = lib.model.Todo.Status.values.map(status => (status.code.toString, status.name)), '_label -> "ステータス")
