+++ conf/routes
@@ -5,18 +5,21 @@
 
 # An example controller showing a sample home page
 GET     /                             controllers.HomeController.index
-GET     /todo                         controllers.HomeController.getTodoList
-POST    /todo                         controllers.HomeController.addTodo
-POST    /todo/edit                    controllers.HomeController.editTodo
-GET     /todo/edit/:id                controllers.HomeController.editTodoHome(id: Int)
-POST    /todo/delete                  controllers.HomeController.deleteTodo(id: Int)
 
-GET     /category                     controllers.CategoryController.getList
-POST    /category                     controllers.CategoryController.add
-GET     /category/add                 controllers.CategoryController.addHome
-GET     /category/edit/:id            controllers.CategoryController.editHome(id: Int)
-POST    /category/edit                controllers.CategoryController.edit
-POST    /category/delete/:id          controllers.CategoryController.delete(id: Int)
+GET     /todo                         controllers.TodoController.getTodoList
+POST    /todo                         controllers.TodoController.addTodo
+POST    /todo/edit                    controllers.TodoController.editTodo
+GET     /todo/:id                     controllers.TodoController.getTodo(id: Int)
+POST    /todo/delete/:id              controllers.TodoController.deleteTodo(id: Int)
+
+GET     /category                     controllers.CategoryController.getCategoryList
+POST    /category                     controllers.CategoryController.addCategory
+GET     /category/add                 controllers.CategoryController.addCategoryHome
+GET     /category/:id                 controllers.CategoryController.getCategory(id: Int)
+POST    /category/edit                controllers.CategoryController.editCategory
+POST    /category/delete/:id          controllers.CategoryController.deleteCategory(id: Int)
+
+
 
 # Map static resources from the /public folder to the /assets URL path
 GET     /assets/*file               controllers.Assets.versioned(path="/public", file: Asset)
